Major refactor of naming. Added functionality for mdd error handling and displaying errror row. General stability improvements

+++ src/dollar/file/dollarfilewriter.py
@@ -1,29 +1,22 @@
 import os
 import shutil
 
-from dollar.confighandler import ConfigHandler
-from dollar.dollarexecutionexception import DollarExecutionException
+from dollar.dollarexception import DollarException
 
 
 class DollarFileWriter:
 
     @staticmethod
-    def write_files(outputs):
-        target_path = ConfigHandler.get("target_path")
-        try:
+    def write_files(target_path, outputs):
+        if os.path.exists(target_path):
             shutil.rmtree(target_path)
-        except:
-            pass # Allow target to not exist
         for output in outputs:
-            save_path = os.path.join(target_path, output.getpath())
+            save_path = os.path.join(target_path, output.get_path())
             save_folder = os.path.split(save_path)[0]
-            try:
+            if not os.path.exists(save_folder):
                 os.makedirs(save_folder)
-            except OSError:
-                # Allow save_folder to exist
-                pass
             try:
                 with open(save_path, "w") as f:
-                    f.write(output.getcontent())
+                    f.write(output.get_content())
             except Exception as e:
-                raise DollarExecutionException("Could not write file {}".format(save_path)) from e
+                raise DollarException("Could not write file {}".format(save_path)) from e
